third_party/blink/web_tests/external/wpt/speech-api/SpeechSynthesisEvent-constructor-expected.txt

This is a testharness.js-based test.
PASS SpeechSynthesisEvent with no arguments throws TypeError
PASS SpeechSynthesisEvent with no eventInitDict throws TypeError
PASS SpeechSynthesisEvent with empty eventInitDict throws TypeError (requires
    utterance)
PASS SpeechSynthesisEvent with eventInitDict not having utterance throws
    TypeError
FAIL SpeechSynthesisEvent with eventInitDict having an utterance assert_equals: expected (number) 0 but got (undefined) undefined
FAIL SpeechSynthesisEvent with custom eventInitDict assert_equals: expected (number) 3 but got (undefined) undefined
Harness: the test ran to completion.
